Create funcs.c
Headers should only contain the definitions of the functions
