+++ funcs.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+
+void print_menu () {
+	printf("Menu\n");
+	printf("1- Log in\n");
+	printf("2- Sign in\n");
+	printf("3- Exit\n");
+}
